moving helper functions to base cli provider
